requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.2
requests==2.31.0
pandas==2.1.3
schedule==1.2.1
pydantic-settings==2.1.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
watchtower==3.0.1
boto3==1.34.11